svn merge -r 40671:41141 https://dart.googlecode.com/svn/branches/bleeding_edge trunk

git-svn-id: http://dart.googlecode.com/svn/trunk@41145 260f80e4-7a28-3924-810f-c04153c831b5

+++ runtime/vm/os_win.cc
@@ -263,7 +263,9 @@ int OS::VSNPrint(char* str, size_t size, const char* format, va_list args) {
   }
   // Make sure to zero-terminate the string if the output was
   // truncated or if there was an error.
-  if (written >= size) {
+  // The static cast is safe here as we have already determined that 'written'
+  // is >= 0.
+  if (static_cast<size_t>(written) >= size) {
     str[size - 1] = '\0';
   }
   return written;
